+++ test_client.py
@@ -6,7 +6,8 @@ echo = servy.client.Client({'name': 'echo', 'host': 'localhost:8000'})
 
 
 def main():
-    # print echo.pars('11')
+    print echo.echo('11')
+    echo.shit()
     return echo.parser.parse_details({'details': 'anything1'})
 
 if __name__ == '__main__':