+++ CHANGELOG.rst
@@ -5,3 +5,5 @@ Changelog
 
 0.2.1
 +++++
+
+-  Add class-based ``Server`` definition